Added admin only requirement to access admin section. Moved unit tests to own directory
