+++ src/app/admin/calipso-user-profile/calipso-user-profile.component.html
@@ -10,6 +10,7 @@
         <h1>User: {{calipsoUser?.user.username}}</h1>
         <h2>Details</h2>
         <h3> Last login: {{calipsoUser?.user.last_login}}</h3>
+        <h3> Admin: {{calipsoUser?.user.is_superuser}}</h3>
       </div>
 
       <div class="col-md-6">
