+++ src/app/tests/admin-nav.component.spec.ts
@@ -1,6 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { AdminNavComponent } from './admin-nav.component';
+import { AdminNavComponent } from '../admin/admin-nav/admin-nav.component';
 
 describe('AdminNavComponent', () => {
   let component: AdminNavComponent;
