+++ src/app/tests/calipso-container-images.component.spec.ts
@@ -1,26 +1,25 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { SelectCalipsoResourceFormComponent } from './select-calipso-resource-form.component';
+import { CalipsoContainerImagesComponent } from '../admin/calipso-container-images/calipso-container-images.component';
 import {HttpClientTestingModule} from '@angular/common/http/testing';
 import {RouterTestingModule} from '@angular/router/testing';
 import {NO_ERRORS_SCHEMA} from '@angular/core';
 
-describe('SelectCalipsoResourceFormComponent', () => {
-  let component: SelectCalipsoResourceFormComponent;
-  let fixture: ComponentFixture<SelectCalipsoResourceFormComponent>;
+describe('CalipsoContainerImagesComponent', () => {
+  let component: CalipsoContainerImagesComponent;
+  let fixture: ComponentFixture<CalipsoContainerImagesComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ SelectCalipsoResourceFormComponent ],
+      declarations: [ CalipsoContainerImagesComponent ],
       imports: [HttpClientTestingModule, RouterTestingModule],
       schemas: [NO_ERRORS_SCHEMA]
-
     })
     .compileComponents();
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(SelectCalipsoResourceFormComponent);
+    fixture = TestBed.createComponent(CalipsoContainerImagesComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
