+++ src/app/tests/calipso-users.component.spec.ts
@@ -1,23 +1,25 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { FacilityNavComponent } from './facility-nav.component';
+import { CalipsoUsersComponent } from '../admin/calipso-users/calipso-users.component';
 import {HttpClientTestingModule} from '@angular/common/http/testing';
 import {RouterTestingModule} from '@angular/router/testing';
+import {NO_ERRORS_SCHEMA} from '@angular/core';
 
-describe('FacilityNavComponent', () => {
-  let component: FacilityNavComponent;
-  let fixture: ComponentFixture<FacilityNavComponent>;
+describe('CalipsoUsersComponent', () => {
+  let component: CalipsoUsersComponent;
+  let fixture: ComponentFixture<CalipsoUsersComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ FacilityNavComponent ],
-      imports: [HttpClientTestingModule, RouterTestingModule]
+      declarations: [ CalipsoUsersComponent ],
+      imports: [HttpClientTestingModule, RouterTestingModule],
+      schemas: [NO_ERRORS_SCHEMA]
     })
     .compileComponents();
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(FacilityNavComponent);
+    fixture = TestBed.createComponent(CalipsoUsersComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
