+++ src/app/tests/calipso-vm.component.spec.ts
@@ -1,6 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { CalipsoVmComponent } from './calipso-vm.component';
+import { CalipsoVmComponent } from '../calipso-vm/calipso-vm.component';
 
 describe('CalipsoVmComponent', () => {
   let component: CalipsoVmComponent;
