+++ src/app/tests/calipsoplus.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed, inject } from '@angular/core/testing';
 
-import { CalipsoplusService } from './calipsoplus.service';
+import { CalipsoplusService } from '../calipsoplus.service';
 import {HttpClientTestingModule} from '@angular/common/http/testing';
 import {RouterTestingModule} from '@angular/router/testing';
