+++ src/app/tests/login-calipso-user-form.component.spec.ts
@@ -1,6 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { LoginCalipsoUserFormComponent } from './login-calipso-user-form.component';
+import { LoginCalipsoUserFormComponent } from '../login-calipso-user-form/login-calipso-user-form.component';
 import {NO_ERRORS_SCHEMA} from '@angular/core';
 import {FormsModule} from '@angular/forms';
 import {HttpClientTestingModule} from '@angular/common/http/testing';
