+++ src/app/tests/select-calipso-quota-form.component.spec.ts
@@ -1,17 +1,17 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { CalipsoVmImagesComponent } from './calipso-vm-images.component';
+import { SelectCalipsoQuotaFormComponent } from '../select-calipso-quota-form/select-calipso-quota-form.component';
 import {HttpClientTestingModule} from '@angular/common/http/testing';
 import {RouterTestingModule} from '@angular/router/testing';
 import {NO_ERRORS_SCHEMA} from '@angular/core';
 
-describe('CalipsoVmImagesComponent', () => {
-  let component: CalipsoVmImagesComponent;
-  let fixture: ComponentFixture<CalipsoVmImagesComponent>;
+describe('SelectCalipsoQuotaFormComponent', () => {
+  let component: SelectCalipsoQuotaFormComponent;
+  let fixture: ComponentFixture<SelectCalipsoQuotaFormComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ CalipsoVmImagesComponent ],
+      declarations: [ SelectCalipsoQuotaFormComponent ],
       imports: [HttpClientTestingModule, RouterTestingModule],
       schemas: [NO_ERRORS_SCHEMA]
     })
@@ -19,7 +19,7 @@ describe('CalipsoVmImagesComponent', () => {
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(CalipsoVmImagesComponent);
+    fixture = TestBed.createComponent(SelectCalipsoQuotaFormComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
