+++ src/app/tests/select-calipso-resource-form.component.spec.ts
@@ -1,25 +1,26 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { AdminDashboardComponent } from './admin-dashboard.component';
+import { SelectCalipsoResourceFormComponent } from '../select-calipso-resource-form/select-calipso-resource-form.component';
 import {HttpClientTestingModule} from '@angular/common/http/testing';
 import {RouterTestingModule} from '@angular/router/testing';
 import {NO_ERRORS_SCHEMA} from '@angular/core';
 
-describe('AdminDashboardComponent', () => {
-  let component: AdminDashboardComponent;
-  let fixture: ComponentFixture<AdminDashboardComponent>;
+describe('SelectCalipsoResourceFormComponent', () => {
+  let component: SelectCalipsoResourceFormComponent;
+  let fixture: ComponentFixture<SelectCalipsoResourceFormComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ AdminDashboardComponent ],
+      declarations: [ SelectCalipsoResourceFormComponent ],
       imports: [HttpClientTestingModule, RouterTestingModule],
       schemas: [NO_ERRORS_SCHEMA]
+
     })
     .compileComponents();
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(AdminDashboardComponent);
+    fixture = TestBed.createComponent(SelectCalipsoResourceFormComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
