+++ src/app/user.ts
@@ -1,5 +1,6 @@
 export class User {
   constructor(
     public username: string,
-    public last_login: string) {  }
+    public last_login: string,
+    public is_superuser: boolean) {  }
 }
